adding collections views
