+++ app/common/directives/modal.js
@@ -5,8 +5,10 @@
 angular.module('ushahidi.common.modal', [])
 
 .directive('modal', [
+    '$translate',
     '$rootScope',
 function (
+    $translate,
     $rootScope
 ) {
     return {
@@ -29,6 +31,7 @@ function (
             $scope.iconPath = '../../img/iconic-sprite.svg#' + $scope.icon;
             $scope.classVisible = false;
             $scope.modalOffset = 0;
+            $scope.title = $translate($scope.title);
 
             // If closeOnOverlayClick isn't passed, default to true
             if (typeof $scope.closeOnOverlayClick === 'undefined') {
