+++ app/common/locales/en.json
@@ -6,9 +6,13 @@
         "can_see_this" : "can see this",
         "collections" : "Collections",
         "create_new" : "Create new",
+        "created_by" : "Created by ",
         "documentation": "Documentation",
         "report_a_bug": "Report a bug",
         "forbidden": "Sorry, you're not allowed to do that.",
+        "get_notifications" : "Get notifications",
+        "turn_off_notifications" : "Turn off notifications",
+        "more" : "More",
         "cancel" : "Cancel",
         "save" : "Save",
         "save_and_close" : "Save &amp; close",
@@ -20,6 +24,7 @@
         "import" : "Import",
         "users" : "Users",
         "edit_user" : "Edit User",
+        "edit" : "Edit",
         "roles" : "Roles",
         "edit_role" : "Edit Role",
         "categories" : "Categories",
