+++ app/set/set-module.js
@@ -6,10 +6,10 @@ angular.module('ushahidi.sets', [])
 .directive('savedsearchEditor', require('./directives/savedsearch-editor-directive.js'))
 .directive('savedsearchCreate', require('./directives/savedsearch-create-directive.js'))
 .directive('savedsearchUpdate', require('./directives/savedsearch-update-directive.js'))
-.directive('collectionEditor', require('./directives/collection-editor-directive.js'))
-.directive('collectionTogglePost', require('./directives/collection-toggle-post.js'))
-.directive('selectCollection', require('./directives/select-collection-directive.js'))
-.directive('collectionListing', require('./directives/collection-listing.js'))
+
+.directive('collectionModeContext', require('./directives/collections/mode-context.js'))
+.directive('collectionEditor', require('./directives/collections/editor.js'))
+.directive('collectionListing', require('./directives/collections/listing.js'))
 
 .service('CollectionEndpoint', require('./services/endpoints/collection.js'))
 .service('SavedSearchEndpoint', require('./services/endpoints/savedsearch.js'));
