+++ app/set/set-routes.js
@@ -6,7 +6,7 @@ function (
     $routeProvider
     .when('/collections/:id/:view?', {
         controller: require('./controllers/collections-controller.js'),
-        templateUrl: 'templates/sets/collections.html',
+        templateUrl: 'templates/sets/collections/collections.html',
         resolve: {
             collection: ['$route', 'CollectionEndpoint', function ($route, CollectionEndpoint) {
