--- server/www/templates/sets/collections.html
@@ -1,19 +1,0 @@
-<div>
-    <mode-context></mode-context>
-
-    <main role="main">
-        <post-view current-view="currentView()" filters="filters" base-url="/collections/{{collection.id}}/">
-            <tool-bar></tool-bar>
-
-            <h1 class="inline">{{collection.name}}</h1>
-            <a ng-if="canEdit()" ng-click="setIsOpen()" class="cta settings" translate>set.edit_collection_settings</a>&nbsp;
-            <a ng-if="showNotificationLink" ng-click="saveNotification(collection)" class="cta settings" translate>set.add_notification</a>
-            <collection-editor collection="collection" is-open="isOpen"></collection-editor>
-            <p class="collection-description">{{collection.description}}</p>
-
-            <div class="applied-filters">
-                <post-active-filters filters="filters"></post-active-filters>
-            </div>
-        </post-view>
-    </main>
-</div>
