+++ server/www/templates/sets/collections/collection-listing.html
@@ -21,7 +21,7 @@
                     </div>
                     <div class="listing-item-primary">
                         <h2 class="listing-item-title">
-                            <a ng-click="collectionClickHandler(collection.id)">
+                            <a ng-click="collectionClickHandler(collection)">
                                 {{collection.name}}
                             </a>
                         </h2>
