+++ server/www/templates/sets/collections/collections.html
@@ -1,0 +1,12 @@
+<div>
+    <collection-mode-context collection="collection"></collection-mode-context>
+
+    <main role="main">
+        <post-view current-view="currentView()" filters="filters" base-url="/collections/{{collection.id}}/">
+            <tool-bar></tool-bar>
+            <div class="applied-filters">
+                <post-active-filters filters="filters"></post-active-filters>
+            </div>
+        </post-view>
+    </main>
+</div>
