+++ server/www/templates/sets/collections/mode-context.html
@@ -1,0 +1,100 @@
+<div class="mode-context init">
+    <header class="mode-context-header">
+        <ol class="breadcrumbs">
+            <li><a href="/">{{site.name}}</a></li>
+        </ol>
+        <h1 class="mode-context-title" translate>{{collection.name}}</h1>
+    </header>
+
+    <span class="mode-context-trigger">
+        <svg class="iconic">
+            <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="../../img/iconic-sprite.svg#chevron-bottom"></use>
+        </svg>
+        <span class="label hidden">app.show_more_less</span>
+    </span>
+    <div class="mode-context-body">
+        <!--// IF: Saved Search has description //-->
+        <p ng-if="collection.description">{{collection.description}}</p>
+        <!--// END IF //-->
+
+        <div class="metadata">
+            <span ng-if="collection.user">
+                <span translate='app.created_by'>Created by</span>
+                <a ng-href="/users/{{collection.user.id}}" class="metadata-author">test{{collection.user.realname}}</a>
+            </span>
+            <span class="metadata-visibility public tooltip">
+                <svg class="iconic">
+                  <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="../../img/iconic-sprite.svg#globe"></use>
+                </svg>
+                <span class="bug">This Collection is visible to the public</span>
+            </span>
+
+            <!--// IF: User has permission to edit the Saved Search //-->
+            <div class="form-field" ng-if="canEditCollection" dropdown>
+                <button type="button" data-modal="edit-collection" ng-click="editCollection()">
+                    <svg class="iconic">
+                        <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="../../img/iconic-sprite.svg#pencil"></use>
+                    </svg>
+                    <span class="button-label" translate>app.edit</span>
+                </button>
+
+                <button type="button" data-toggle="dropdown-menu" class="init" dropdown-toggle>
+                    <svg class="iconic">
+                        <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="../../img/iconic-sprite.svg#ellipses"></use>
+                    </svg>
+                    <span class="button-label hidden" translate>app.more</span>
+                </button>
+                <ul class="dropdown-menu init" dropdown-menu>
+                    <li ng-if="showNotificationLink">
+                        <a ng-click="saveNotification(collection)">
+                            <svg class="iconic">
+                                <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="../../img/iconic-sprite.svg#bell"></use>
+                            </svg>
+                            <span class="label" ng-click="saveNotification(collection)" translate="app.get_notifications">Get notifications</span>
+                        </a>
+                    </li>
+                    <li ng-if="!showNotificationLink">
+                        <a ng-click="removeNotification(collection)">
+                            <svg class="iconic">
+                                <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="../../img/iconic-sprite.svg#bell"></use>
+                            </svg>
+                            <span class="label" translate="app.turn_off_notifications">Turn off notifications</span>
+                        </a>
+                    </li>
+                    <li>
+                        <a ng-click="deleteCollection()">
+                            <svg class="iconic">
+                                <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="../../img/iconic-sprite.svg#trash"></use>
+                            </svg>
+                            <span class="label" translate="app.delete"></span>
+                        </a>
+                    </li>
+                </ul>
+            </div>
+            <!--// END IF //-->
+
+            <!--// IF: User does not have permission to edit the Saved Search //-->
+            <div class="form-field" ng-if="!canEditCollection">
+                <!-- // IF: User is not yet getting notifications about this Saved Search //-->
+                <button type="button" ng-if="showNotificationLink" ng-click="saveNotification(collection)">
+                    <svg class="iconic">
+                        <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="../../img/iconic-sprite.svg#bell"></use>
+                    </svg>
+                    <span class="button-label" translate="app.get_notifications">Get notifications</span>
+                </button>
+
+                <!--// ELSE //-->
+
+                <button type="button" ng-if="!showNotificationLink" class="button-flat">
+                    <svg class="iconic">
+                        <use xmlns:xlink="http://www.w3.org/1999/xlink" xlink:href="../../img/iconic-sprite.svg#bell"></use>
+                    </svg>
+                    <span class="button-label" translate="app.turn_off_notifications">Turn off notifications</span>
+                </button>
+                <!--// END IF //-->
+            </div>
+            <!--// END IF //-->
+        </div>
+
+    </div>
+</div>
