--- server/www/templates/views/select-collection.html
@@ -1,6 +1,0 @@
-<div ng-show="collections.length > 0" class="custom-select">
-  <select ng-model="selectedCollectionId">
-    <option value="" translate>set.add_to_collection</option>
-    <option value="{{collection.id}}" ng-repeat="collection in collections">{{collection.name}}</option>
-  </select>
-</div>
